// types.ts
export type BrowsersListBrowserName =
  | "ie"
  | "edge"
  | "firefox"
  | "chrome"
  | "safari"
  | "opera"
  | "ios_saf"
  | "op_mini"
  | "android"
  | "bb"
  | "op_mob"
  | "and_chr"
  | "and_ff"
  | "ie_mob"
  | "and_uc"
  | "samsung"
  | "and_qq"
  | "baidu"
  | "kaios"
  | "fx"
  | "ff"
  | "ios"
  | "explorer"
  | "blackberry"
  | "explorermobile"
  | "operamini"
  | "operamobile"
  | "chromeandroid"
  | "firefoxandroid"
  | "ucandroid"
  | "qqandroid";